+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Other
 
-* Archetype access speedup by 10%, by using a fixed-size array of storage column references, and elimination of bounds checks (!125)
+* Speedup archetype access by 10%, by elimination of bounds checks (!125)
 
 ## [[v0.4.5]](https://github.com/mlange-42/arche/compare/v0.4.4...v0.4.5)